Move H2DatabaseConfig to its own file.
