+++ src/main/scala/co/teapot/tempest/server/DatabaseConfig.scala
@@ -1,19 +1,7 @@
 package co.teapot.tempest.server
-import com.zaxxer.hikari.{HikariConfig, HikariDataSource}
 
-/**
-  * Created by gkk on 6/14/17.
-  */
+import com.zaxxer.hikari.HikariDataSource
+
 trait DatabaseConfig {
-
   def createConnectionSource(): HikariDataSource
-
-}
-
-class H2DatabaseConfig extends DatabaseConfig {
-  lazy val createConnectionSource: HikariDataSource = {
-    val config = new HikariConfig()
-    config.setJdbcUrl("jdbc:h2:mem:test")
-    new HikariDataSource(config)
-  }
 }
