+++ src/main/scala/co/teapot/tempest/server/H2DatabaseConfig.scala
@@ -1,0 +1,11 @@
+package co.teapot.tempest.server
+
+import com.zaxxer.hikari.{HikariConfig, HikariDataSource}
+
+class H2DatabaseConfig extends DatabaseConfig {
+  lazy val createConnectionSource: HikariDataSource = {
+    val config = new HikariConfig()
+    config.setJdbcUrl("jdbc:h2:mem:test")
+    new HikariDataSource(config)
+  }
+}
